+++ react-container/src/__mocks/auth-utils.ts
@@ -1,0 +1,21 @@
+export const mock = {
+  getFlow: () => {
+    return {
+      flow: 'test-flow',
+    };
+  },
+  CogniteAuth: (client: unknown) => ({
+    state: {},
+    getClient: () => client,
+    onAuthChanged: (
+      applicationId: string,
+      callback: (state: unknown) => void
+    ) => {
+      callback({
+        applicationId,
+        authenticated: true,
+      });
+      return jest.fn();
+    },
+  }),
+};