+++ react-container/src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './AuthContainer';
-export * from './CdfAuthContainer';
 export * from './TranslationWrapper';