+++ react-container/src/utils/index.ts
@@ -1,7 +1,6 @@
 export * from './log';
 export * from './env';
 export * from './localStorage';
-export * from './getSDKClient';
 export * from './tenancy';
 export * from './sidecar';
 export { retrieveAccessToken } from './persistance';